requirements.txt
streamlit>=1.30.0
openai>=1.10.0
numpy>=1.24.0
sentence-transformers>=2.2.0
torch>=2.0.0
PyMuPDF>=1.23.0

# Persistence Layer
pymongo>=4.6.0
faiss-cpu>=1.7.0

# Advanced Extraction Pipeline
pdfplumber>=0.10.0
pdf2image>=1.16.0
Pillow>=10.0.0
pytesseract>=0.3.10
networkx>=3.0
tabulate>=0.9.0

# Observability (Langfuse)
# Pin strictly to v2.x: SDK v3.0+ AND v4.x both switched to OpenTelemetry (OTLP)
# transport, which requires Langfuse server v3+. The docker-compose ships
# langfuse/langfuse:2 (server v2) which only exposes /api/public/ingestion (REST).
# SDK v2.x uses pure REST and is fully compatible with server v2.
#
# To unlock SDK v3+/v4: upgrade the server first:
#   docker-compose-langfuse.yml → image: langfuse/langfuse:3  (new web+worker arch)
#   Then change this pin to:  langfuse>=3.0.0,<5.0.0
langfuse>=2.0.0,<3.0.0
